Support :ns-tracker/resource-deps metadata
Some libraries such as HugSQL define functions in a namespace based on
the contents of a static resource file. When that file is changed, the
namespace needs to be reloaded to reflect the changes.

This change adds support for a :ns-tracker/resource-deps metadata in the
namespace declaration for listing the resources that the namespace
depends on, to facilitate reloading the namespace when one of the
resources changes.

Resolves: #23